refactor: add hero component to responsively update
- render different component based on screensize
- either render carousel or image grid
