+++ src/App.tsx
@@ -3,8 +3,8 @@ import { BrowserRouter as Router } from 'react-router-dom';
 import GlobalStyle from './App.styles';
 import Activities from './components/Activities/Activities';
 import Header from './components/Header/Header';
-import Hero from './components/Hero/Hero';
 import Navbar from './components/shared/Navbar/Navbar';
+import Hero from './components/Hero/Hero';
 
 function App() {
   return (
