+++ src/components/shared/ImageGrid/ImageGrid.styles.ts
